TGDS 2026:  -TGDS-MB v3 is fully integrated on TGDS + WoopsiSDK project.
Stable on NTR/TWL hardware and guaranteed to work on such hardware (DSiXL as well a DS Phat).

-Fixed TGDS NTR/TWL compatibility in popular loaders.

-

TGDS 2025:
-Jan 31, 2025: Implement proper ARM7(i) and ARM9(i) IntrWait/VBlankIntrWait bios calls replacement.
Now TWL mode works correctly on stuff like TGDSProject3D

-

Update int runThreads(struct task_Context * taskCtx, bool waitForVblk), now CPU sleeps until any interrupt, or a vblank interrupt.

-

-Add getTGDSThreadSystem(), pauseThread(), resumeThread(), removeThread().
-ARM7 as well ARM9 uses TGDS threading now.
-Moved ARM7 touchscreen code as task.
-Threading is initialized by default, TGDS Projects will just require to inject tasks to run them.

-

-wait for irq when tasks are done, to prevent artificial locksteps between threads.

Note: This commit removes unnecessary stuttering on TGDS-multimediaplayer.

-

-Threads are ran independently over milliseconds or microseconds (highest priority).

This enables a TGDS Project to save timers, keep CPU usage low, being highly performant as well run threads on a timely basis.

-

Update TGDS Threading:

tUnitsMilliseconds(Milliseconds) are queued and counted up to NDS timers. Assigned to less resource intensive tasks. (CPU + Timer interrupts)
tUnitsMicroseconds(Microseconds) are given the highest thread priority. Assigned to most resource intensive tasks (CPU interrupts)

-

-increase threading granularity, so we can count between milliseconds or microseconds, to give tasks more CPU cycles

-

Restore GDB Stub support.
If an Exception occurs, and TGDS Project has WIFI available for ARM7 & ARM9 cores, GDB debugging will be enabled. Otherwise, it's disabled.

-

TGDS Project now uses threading for process:

-handleTurnOnTurnOffScreenTimeout();

-

-add -fno-short-enums, fixes erratic code behaviour on TGDS Filesystem

-

8 June 2025: Migrate TGDS ecosystem from Bitbucket into Github

+++ arm9/source/TGDSMemoryAllocator/xmem.c
@@ -12,9 +12,10 @@
 #include <string.h>
 #include "typedefsTGDS.h"
 #include "xmem.h"
-#include "utilsTGDS.h"
+#include "posixHandleTGDS.h"
+#include "InterruptsARMCores_h.h"
 
-// default use 1.5 MB
+// default use 128K (ARM9 Mapped), may be overriden.
 unsigned int XMEMTOTALSIZE = (128*1024);
 
 // how many bytes will each of our blocks be?
@@ -30,6 +31,7 @@ unsigned int XMEM_TABLESIZE = 0;
 #define XMEM_ENDBLOCK 0x02
 #define XMEM_USEDBLOCK 0x04
 
+
 unsigned char *xmem_table;
 //XMEM_BLOCK *xmem_blocks;
 unsigned char *xmem_blocks;
@@ -42,8 +44,10 @@ __attribute__((optimize("O0")))
 __attribute__ ((optnone))
 #endif
 void XmemSetup(unsigned int size, unsigned short blocks) {
+
 	XMEMTOTALSIZE = size;
 	XMEM_BLOCKSIZE = blocks;
+	
 }
 
 #if (defined(__GNUC__) && !defined(__clang__))
@@ -191,11 +195,11 @@ void Xfree(const void *ptr) {
 	int block,sblock;
 	
 	while (1) {
-		if (ptr < xmem_blocks) {
+		if ((int)ptr < (int)xmem_blocks) {
 			//printf("XM: Free: NXML %8.8X ",(unsigned int)ptr);
 			break;
 		}
-		if (ptr > (xmem_blocks+(XMEM_BLOCKCOUNT*XMEM_BLOCKSIZE))) {
+		if ((int)ptr > ((int)(xmem_blocks+(XMEM_BLOCKCOUNT*XMEM_BLOCKSIZE))) ) {
 			//printf("XM: Free: NXMG %8.8X ",(u32)ptr);
 			break;
 		}
